requirements.txt
blinker==1.7.0
build==1.1.1
click==8.1.7
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
packaging==24.0
PyMySQL==1.1.0
pyproject_hooks==1.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.28
tomli==2.0.1
typing_extensions==4.10.0
Werkzeug==3.0.1
